refactor: use only uint64 for currencies in the backend

+++ cmd/api/controllers/receipts.go
@@ -94,7 +94,7 @@ func (d *ReceiptsController) ListReceipts(c *gin.Context) {
 			Date:          r.Date.AsTime().Format("2006-01-02"),
 			Vendor:        strings.Title(r.Vendor),
 			PendingReview: pendingReview,
-			TotalAmount:   fmt.Sprintf("%0.2f", mcduck.ConvertToDollar(int32(r.TotalAmount))),
+			TotalAmount:   fmt.Sprintf("%0.2f", mcduck.ConvertToDollar(r.TotalAmount)),
 		}
 
 		viewModels = append(viewModels, v)
